requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
python-dotenv==1.0.1
openai
requests
google-generativeai==0.8.6
groq==1.1.1
xhtml2pdf==0.2.17
beautifulsoup4==4.12.3g
google-auth==2.29.0